STM32 WATCHDOG : compilation issue with typed define

+++ targets/TARGET_STM/TARGET_STM32L0/device/stm32l0xx_hal_conf.h
@@ -132,7 +132,7 @@
   * @brief Internal Low Speed oscillator (LSI) value.
   */
 #if !defined  (LSI_VALUE)
- #define LSI_VALUE  ((uint32_t)37000U)       /*!< LSI Typical Value in Hz*/
+ #define LSI_VALUE  (37000U)       /*!< LSI Typical Value in Hz*/
 #endif /* LSI_VALUE */                      /*!< Value of the Internal Low Speed oscillator in Hz
                                              The real value may vary depending on the variations
                                              in voltage and temperature.*/
